Fix bug in previous patch for avoiding notifying for initial data
Now we will instead return a specially constructed empty set that
can be used to distinguish empty result from missing result.

There is a new method in ChangeNotifers that consumers can use to
detect this.

I apologize for the hacky code, if the major version is ever
bumped this should be revert so that the API can return an
Optional<Set<T>> instead, or avoid firing in ChangeNotifier.setListener(listener, true)
if no change has ever happened